Fix logic of GHIDRATHON_SAVE_PATH value
The following commit fixes incorrect login in the previous commit.
Please refer to the explination below:

if GHIDRATHON_SAVE_PATH is set to any value (other than None or ""),
check if it exists and is a directory.

elif GHIDRATHON_SAVE_PATH is an empty string, report an invalid
directory to avoid confusion.

else GHIDRATHON_SAVE_PATH is not, use default install path.